requirements/base.txt
django
djangorestframework
django-environ
django-filter
bootstrap-admin
drf-yasg
swagger_spec_validator
flex
pycrypto
